av1: CB2 fix av1 build error on buildroot. [1/1]
PD#SWPL-75867

Problem:
buildroot patchbuild failed.

Solution:
fix the wrong format string in print code.

Verify:
buildroot patchbuild.

+++ drivers/frame_provider/decoder_v4l/vav1/av1_film_grain.c
@@ -1136,8 +1136,8 @@ void av1_add_film_grain_run(char *fg_table_buf, struct aom_film_grain_t *params)
 
 	if (debug_fgs & DEBUG_FGS_CONSUME_TIME) {
 		time6 = div64_u64(local_clock() - step_time, 1000);
-		pr_info("fgs consume time %d (%d, %d, %d, %d, %d, %d)us\n",
-			div64_u64(local_clock() - start_time, 1000), time1, time2, time3, time4, time5, time6);
+		pr_info("fgs consume time %ld (%d, %d, %d, %d, %d, %d)us\n",
+			(ulong)(div64_u64(local_clock() - start_time, 1000)), time1, time2, time3, time4, time5, time6);
 	}
 }
